Fix removed frameset

+++ branches/mgallery/webEdition/liveUpdate/sqldumps/tblWorkflowDocTask.sql
@@ -4,6 +4,6 @@ CREATE TABLE ###TBLPREFIX###tblWorkflowDocTask (
   workflowTaskID int(11) unsigned NOT NULL default '0',
   `Date` int(10) unsigned NOT NULL default '0',
   todoID int(11) unsigned NOT NULL default '0',
-  `Status` int(11) unsigned NOT NULL default '0',
+  `Status` tinyint(1) unsigned NOT NULL default '0',
   PRIMARY KEY  (ID)
 ) ENGINE=MyISAM;
